First workign(unstable) version

+++ Diplomski V2.1/ControllerDriverRPI/globals.h
@@ -10,7 +10,7 @@ typedef struct Message {
 	int keyCode;
 	int scanCode;
 	int size;
-	int* content;
+	uint8_t* content;
 }Message;
 
 extern int clientSocket;
